requirements.txt
# Core data processing and API interaction
requests>=2.25.0,<3.0.0

# Environment variable management (for .env files)
python-dotenv>=0.19.0,<1.1.0

# LangChain - Core components, OpenAI integration, ChromaDB vector store, text splitters
langchain>=0.1.0,<0.2.0
langchain-core>=0.1.15,<0.2.0
langchain-community>=0.0.15,<0.1.0
langchain-openai>=0.0.1,<0.1.0 

# OpenAI SDK (often a dependency of langchain-openai, but good to specify)
openai>=1.0.0,<2.0.0

# Tiktoken (used by OpenAI models for tokenization, often a dependency)
tiktoken>=0.5.0,<0.6.0

# ChromaDB - Vector database client library
chromadb>=0.4.20,<0.5.0

# Optional: For more advanced text processing if you extend chunk_text.py
# nltk>=3.6.0,<4.0.0 
# sentence-transformers>=2.2.0,<3.0.0 # If you decide to use local Hugging Face embedding models

# Logging is part of the Python standard library, no explicit install needed.
# json is part of the Python standard library.
# os is part of the Python standard library.
# time is part of the Python standard library.
# uuid is part of the Python standard library.
# datetime is part of the Python standard library.